index on 2.6.0-dev: 08cdee9 Merge pull request #26 from enyojs/GT-16924-derektor
