+++ test/moonstone/TimePicker/GT-11132-Reset-specs.js
@@ -2,7 +2,7 @@ var helpers = rootRequire('./helpers'),
 	app = {};	// Test-specific settings at bottom of the file
 
 var base = 'http://localhost:3000/',
-	url = 'ui-tests/test/loader.html?moonstone/TimePicker/GT-11132-Reset',
+	path = 'test/moonstone/TimePicker/GT-11132-Reset',
 	title = 'Time Picker: Reset Time',
 	tags = ['moonstone', 'qa', 'TimePicker'];	// Tags show up in SauceLabs test output
 
@@ -10,6 +10,7 @@ describe(title, function() {
 	var browser;
 
 	before(function(done) {
+		helpers.epack(path);
 		browser = helpers.initBrowser(title, tags, base, done);
 	});
 
@@ -22,7 +23,7 @@ describe(title, function() {
 	it('should clear time when time set to null', function (done) {
 		browser
 			.setWindowSize(1920,1280)
-			.get(url)
+			.get(base + 'ui-tests/dist')
 			.waitForElementById(app.pickerID)
 			.enyoPropertyGet(app.pickerID, 'value').should.eventually.equal(null, 'starts empty')
 			.elementById(app.pickerID)
@@ -40,7 +41,7 @@ describe(title, function() {
 });
 
 app = {
-	pickerID: 'app_pickerTime',
-	resetID: 'app_buttonReset'
+	pickerID: 'gT-11132-Reset_pickerTime',
+	resetID: 'gT-11132-Reset_buttonReset'
 };
