+++ test/moonstone/TimePicker/GT-11132-Reset/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "GT-11132-Reset.js"
+}
